refactor: initialize slash commands on ready
