+++ .github/workflows/gce.yaml
@@ -18,6 +18,7 @@ on:
   push:
     branches:
     - 'master'
+    - 'slash-commands'
 
 env:
   PROJECT_ID: ${{ secrets.GCE_PROJECT }}
